Refactored functionality and removed dead classes
